added an api method to check if a player is casting

+++ src/main/java/com/runicrealms/plugin/spellapi/spelltypes/ISpell.java
@@ -30,6 +30,6 @@ public interface ISpell {
 
     boolean verifyEnemy(Player caster, Entity victim); // check tons of things, like if target entity is NPC, party member, and outlaw checks
 
-    void addStatusEffect(Entity entity, EffectEnum effectEnum, int duration);
+    void addStatusEffect(Entity entity, EffectEnum effectEnum, double duration);
 }
